chore(influx): drop REPL command
closes: #18676
